% Auto-generated: do not edit by hand
\name{dashPaperdragon}

\alias{dashPaperdragon}

\title{DashPaperdragon component}

\description{

}

\usage{
dashPaperdragon(id=NULL, curMousePosition=NULL, curShapeObject=NULL,
imageSrc=NULL, shapeList=NULL, viewPortBounds=NULL,
zoomLevel=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{curMousePosition}{Named list. Current Mouse Position in Image Coordinates}

\item{curShapeObject}{Named list. curShapeObject is the current shape object that was most recently moused over}

\item{imageSrc}{Character. the tile source for openseadragon}

\item{shapeList}{Named list. shapeList is a list of shapes to be drawn on the image}

\item{viewPortBounds}{Named list. viewportBounds of the current OSD Viewer}

\item{zoomLevel}{Numeric. zoomLevel of the current OSD Viewer}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

