% Auto-generated: do not edit by hand
\name{dashPaperdragon}

\alias{dashPaperdragon}

\title{DashPaperdragon component}

\description{

}

\usage{
dashPaperdragon(id=NULL, config=NULL, curMousePosition=NULL, imageSrc=NULL,
inputToPaper=NULL, outputFromPaper=NULL,
viewportBounds=NULL, zoomLevel=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{config}{Named list. config is a dictionary of options for setting up the component and initial rendering}

\item{curMousePosition}{Named list. Current Mouse Position in Image Coordinates}

\item{imageSrc}{Character. the tile source for openseadragon}

\item{inputToPaper}{Named list. data sent from dash to paper}

\item{outputFromPaper}{Named list. data sent from paper back to dash}

\item{viewportBounds}{Named list. viewportBounds of the current OSD Viewer}

\item{zoomLevel}{Numeric. zoomLevel of the current OSD Viewer}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

